+++ automated-orders/packages/blockchain-library/src/index.ts
@@ -4,3 +4,23 @@ export { NewSignatureProcessor } from './processors/newSignatureProcessor.js';
 export { OrderProcessor } from './processors/orderProcessor.js';
 export { INewSignatureRequest } from './types/warden/newSignatureRequest.js';
 export { IOrderRegistered, OrderRegisteredAbi } from './types/order/events.js';
+export {
+    SignRequestStatus,
+    BroadcastType,
+    ISignRequest, 
+    IPageRequest,
+    ISignRequestResponse, 
+    SignRequestsAbi,
+} from './types/warden/functions.js';
+
+export { WardenClient } from './clients/warden.js';
+
+export { WardenRegistryClient } from './clients/registry.js';
+export { IWardenRegistryConfiguration } from './types/registry/configuration.js';
+
+export {
+    AddTransactionAbi ,
+    ExecutionsAbi ,
+    RegisterAbi,
+    TransactionsAbi,
+} from './types/registry/functions.js';