+++ automated-orders/packages/relayer/src/config/config.ts
@@ -4,11 +4,21 @@ import { Env } from './env.js';
 
 const schema: JSONSchemaType<Env> = {
   type: 'object',
-  required: ['ETHEREUM_NODE_RPC', 'WARDEN_RPC_URL', 'WARDEN_POLLING_INTERVAL_MSEC'],
+  required: [
+    'EVMOS_NODE_RPC',
+    'EVMOS_REGISTRY_ADDRESS',
+    'ETHEREUM_NODE_RPC',
+    'WARDEN_EVM_CHAIN_ID', 
+    'WARDEN_SIGN_REQUESTS_POLLING_INTERVAL_MSEC',
+    'WARDEN_SIGN_REQUESTS_PAGE_SIZE',
+  ],
   properties: {
+    EVMOS_NODE_RPC: { type: 'string' },
+    EVMOS_REGISTRY_ADDRESS: { type: 'string' },
     ETHEREUM_NODE_RPC: { type: 'string' },
-    WARDEN_RPC_URL: { type: 'string' },
-    WARDEN_POLLING_INTERVAL_MSEC: { type: 'number' },
+    WARDEN_EVM_CHAIN_ID: { type: 'number' },
+    WARDEN_SIGN_REQUESTS_POLLING_INTERVAL_MSEC: { type: 'number', minimum: 100 },
+    WARDEN_SIGN_REQUESTS_PAGE_SIZE: { type: 'number', minimum: 1 },
   },
 };