+++ solidity/orders/test/BasicOrder.t.sol
@@ -65,9 +65,8 @@ contract BasicOrderTest is Test {
     // @openzeppelin Ownable.sol
     error OwnableUnauthorizedAccount(address account);
 
-    address private constant SEPOLIA_UNISWAP_V2_ROUTER =
-        address(bytes20(bytes("0xeE567Fe1712Faf6149d80dA1E6934E354124CfE3")));
-    address private constant RECEIVER = address(bytes20(bytes("0x18517Cb2779186B86b1F8947dFdB6078C1B9C9db")));
+    address private constant SEPOLIA_UNISWAP_V2_ROUTER = address(uint160(0xeE567Fe1712Faf6149d80dA1E6934E354124CfE3));
+    address private constant RECEIVER = address(uint160(0x18517Cb2779186B86b1F8947dFdB6078C1B9C9db));
 
     function beforeTestSetup(bytes4 testSelector) public pure returns (bytes[] memory beforeTestCalldata) {
         if (