+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * (precompiles) Add an ability for contracts to approve actions
 * (go-client) Return transaction hash from SendWaitTx
 * (keychain-sdk) Log transaction hashes of broadcasted transactions
+* (relayer) #992 Added a service to relay transactions into Ethereum
 
 ### Consensus Breaking Changes
 * (precompiles) Add slinky precompiled contract